renamed input class
