+++ src/main/java/uk/gov/cslearning/record/service/CourseRecordService.java
@@ -1,18 +1,15 @@
 package uk.gov.cslearning.record.service;
 
 import com.github.fge.jsonpatch.JsonPatch;
-import com.github.fge.jsonpatch.JsonPatchException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
-import org.springframework.web.client.HttpClientErrorException;
-import uk.gov.cslearning.record.api.input.CourseRecordInput;
+
+import uk.gov.cslearning.record.api.input.PATCH.PatchCourseRecordInput;
 import uk.gov.cslearning.record.api.mapper.CourseRecordMapper;
 import uk.gov.cslearning.record.api.util.PatchHelper;
 import uk.gov.cslearning.record.domain.CourseRecord;
 import uk.gov.cslearning.record.exception.CourseRecordNotFoundException;
-import uk.gov.cslearning.record.exception.PatchResourceException;
 import uk.gov.cslearning.record.repository.CourseRecordRepository;
 
 @Service
@@ -25,9 +22,9 @@ public class CourseRecordService {
 
     public CourseRecord updateCourseRecord(String userId, String courseId, JsonPatch patch) {
         CourseRecord courseRecord = courseRecordRepository.getCourseRecord(userId, courseId).orElseThrow(() -> new CourseRecordNotFoundException(userId, courseId));
-        CourseRecordInput updateParams = courseRecordMapper.asInput(courseRecord);
+        PatchCourseRecordInput updateParams = courseRecordMapper.asInput(courseRecord);
 
-        CourseRecordInput patchedInput =  patchHelper.patch(patch, updateParams, CourseRecordInput.class);
+        PatchCourseRecordInput patchedInput =  patchHelper.patch(patch, updateParams, PatchCourseRecordInput.class);
         courseRecordMapper.update(courseRecord, patchedInput);
 
         courseRecord.setLastUpdated(LocalDateTime.now());
